+++ tests/testthat/test-receptiviti_status.R
@@ -2,6 +2,7 @@ test_that("failures works", {
   expect_error(receptiviti_status(key = ""), "specify your key")
   expect_error(receptiviti_status(key = 123, secret = ""), "specify your secret")
   expect_error(receptiviti_status("localhost", key = 123, secret = 123), "url does not appear to be valid")
+  expect_null(receptiviti_status("http://localhost:0/not_served", key = 123, secret = 123))
   expect_identical(
     capture.output(receptiviti_status("example.com", key = 123, secret = 123), type = "message")[2],
     "Message: 404"